clearing posts - users relations

+++ REST_API/controllers/feed.js
@@ -115,6 +115,12 @@ exports.updatePost = (req, res, next) => {
             error.statusCode = 404;
             throw error
         }
+        // Authorization
+        if (post.creator.toString() !== req.userId) {
+            const error = new Error('Not authorized');
+            error.statusCode = 403;
+            throw error
+        }
         if(imageUrl !== post.imageUrl) {
             clearImage(post.imageUrl)
         }
@@ -146,9 +152,20 @@ exports.deletePost = (req, res, next) => {
             error.statusCode = 404;
             throw error
         }
+        // Authorization
+        if (post.creator.toString() !== req.userId) {
+            const error = new Error('Not authorized');
+            error.statusCode = 403;
+            throw error
+        }
         // clear image
         clearImage(post.imageUrl)
         return Post.findByIdAndRemove(postId)
+    }).then(result => {
+        return User.findById(req.userId)
+    }).then(user => {
+        user.posts.pull(postId);
+        return user.save()
     }).then(result => {
         console.log(result);
         res.status(200).json({ message: 'Deleted post!' })
